refactor: split into modules, and added psr4 autoload
