+++ composer.json
@@ -1,0 +1,13 @@
+{
+    "name": "jorgen/netex-converter",
+    "description": "Convert NeTEx ZIP to GeoJSON and optionally MBTiles",
+    "type": "project",
+    "require": {
+        "php": ">=8.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "App\\": "src/"
+        }
+    }
+}
